feat:(server) optimize `items_crafted` table by deleting entries of characters that don't exist anymore
this will run at every server restart
the idea was to add it to the table a constrain  , but that would only work for newer servers.
I found it easy to do it this way instead of telling users to add constrains (not all will do or miss it reading it), this way we force it.