fix: update labels descriptions and serial numbers for clients when exports are used
only when dropping or giving item would update when the exports to set label desc or serial number  when these were used

so now we check in client who has the weapon id to update it, instead of a callback each time to get info updated when it will only be used once most likely.
there's many ways to do this i find it easier with less checks to just send to all clients who has the weapon it will update when the export is used.