fix(server): addItemsToCustomInventory
since we miss some data on the database we can update the table with the items already in the container this only works when opening the inventory

to use the export we need to check if item is in memory if not we add. this way we can update cache

eventually would be good to add to database if inventory is shared and identifiers so we can load it on server start instead of every time we try to update an item that already exists.